+++ client_src/js/exceptions/application-error.handler.js
@@ -1,0 +1,25 @@
+import { errorTypes } from "./error.types";
+
+export class ApplicationError extends Error {
+  static hasErrorNotificationBeenTriggered = false;
+  constructor(options, overrides) {
+    super();
+    Object.assign(options, overrides);
+
+    this.name = "OctoFarm Error";
+    this.type = options.type;
+    this.code = options.code;
+    this.message = options.message;
+    this.errors = options.errors;
+    this.meta = options.meta;
+    this.statusCode = options.statusCode;
+    this.color = options.color;
+    // {
+    //   analytics:  {},
+    //   context: {}
+    // }
+
+    // Show error message to client...
+    ApplicationError.hasErrorNotificationBeenTriggered = true;
+  }
+}