+++ client_src/js/historyRunner.js
@@ -2,7 +2,7 @@ import Calc from "./lib/functions/calc.js";
 import UI from "./lib/functions/ui.js";
 import { returnDropDown } from "./lib/modules/filamentGrab.js";
 import * as ApexCharts from "apexcharts";
-import OctoFarmClient from "./lib/octofarm_client";
+import OctoFarmClient from "./services/octofarm_client.service";
 
 // Setup history listeners
 document.getElementById("historyTable").addEventListener("click", (e) => {