+++ client_src/js/list-view.runner.js
@@ -1,5 +1,5 @@
 import { setViewType } from "./monitoring/monitoring-view.state";
-import { createClientSSEWorker } from "./lib/client-worker";
+import { createClientSSEWorker } from "./services/client-worker.service";
 import { monitoringSSEventHandler, monitoringWorkerURL } from "./monitoring/monitoring-sse.handler";
 
 const viewType = "list";